Added new generateChartDatasets()
Merged the functions simplePie() and memes() into one function, generateChartDataset(). This function is designed to provide the same functionality of both functions simplePie() and memes(), it will accept and array of strings representing the columns that it will be rendering into a single pie or bar chart. Some string arrays from poll7.js have also been moved into charts.json in an attempt to further abstract the [poll].js component