First Utilization of [poll].js
Moved over all of the code making up poll7 into the dynamically router [poll].js page. This page takes in a number in the URL and uses it to specify a poll to render data for. There is also a small fix in poll7.js, somehow the gernman 'ss' character found its way in the code, not sure how that got there, and its not the first time this has happend