added support for piecharts that combine datapoints
Added a check for a 'shape' key within each charts dataset. The 'shape' value is an array containing more arrays that dictate how the datapoints should be grouped together. For example, Perhaps a survery question has 10 possible responses, but we would like to somehow combine them to form 3 datapoints and render them in a chart. Within each array as an integer value which is supposed to represent the indicies and that those indicies values should be group together into one value.