Finished abstracting hardcoded chartconfigs
Created a function, combineDataPoints(), that takes in a chart object and an array of genereated datasets to be fed into charts. Doing it in this way allows for chart objects to be read, created, and generated without needing to be named or predefined in the [poll].js component. The next later to this paradigm is to create a loop that will iterated through the array of chart objects and run combineDataPoints on each chart object.