finalized abstracting the Poll page
After these changes, the dynamically routed next.js [poll].js page component should now be able to handle any array of chart objects, in any order/type, and render them nicely in a CSS grid container. It can handle one level of nested chartGroup chart objects. This component requires an array of chart objects to be passed in as a .json file, which it will then read from and perform operations to render the charts defined inside of that .json file