Updated side nav bar for home page to push content
I used the css grid to make a layout for the home page in index.js, in order add content later. Modified the sidebar navigation properties so that it could be displayed as soon as the page was loaded without having to click on a button, and made sure it pushed all content to the right.